requirements.txt
Pillow
fastapi
uvicorn
numpy
tensorflow
python-multipart